Enhances update process and adds version info
Improves the update process for both the main script and the base Docker image.

Now records and displays the last update time and commit version for better user awareness.

Also updates the base repo `docker-minecraft-server` to ensure the user is working with the latest version, separating this from the main script updates.